+++ cvat-ui/src/reducers/interfaces.ts
@@ -13,6 +13,10 @@ export interface AuthState {
     initialized: boolean;
     fetching: boolean;
     user: any;
+    authActionsFetching: boolean;
+    authActionsInitialized: boolean;
+    showChangePasswordDialog: boolean;
+    allowChangePassword: boolean;
 }
 
 export interface TasksQuery {
@@ -176,6 +180,8 @@ export interface NotificationsState {
             login: null | ErrorState;
             logout: null | ErrorState;
             register: null | ErrorState;
+            changePassword: null | ErrorState;
+            loadAuthActions: null | ErrorState;
         };
         tasks: {
             fetching: null | ErrorState;
@@ -245,6 +251,9 @@ export interface NotificationsState {
         };
         models: {
             inferenceDone: string;
+        };
+        auth: {
+            changePasswordDone: string;
         };
     };
 }