+++ cvat-ui/src/reducers/notifications-reducer.ts
@@ -27,6 +27,8 @@ const defaultState: NotificationsState = {
             login: null,
             logout: null,
             register: null,
+            changePassword: null,
+            loadAuthActions: null,
         },
         tasks: {
             fetching: null,
@@ -97,6 +99,9 @@ const defaultState: NotificationsState = {
         models: {
             inferenceDone: '',
         },
+        auth: {
+            changePasswordDone: '',
+        },
     },
 };
 
@@ -156,6 +161,48 @@ export default function (state = defaultState, action: AnyAction): Notifications
                         ...state.errors.auth,
                         register: {
                             message: 'Could not register on the server',
+                            reason: action.payload.error.toString(),
+                        },
+                    },
+                },
+            };
+        }
+        case AuthActionTypes.CHANGE_PASSWORD_SUCCESS: {
+            return {
+                ...state,
+                messages: {
+                    ...state.messages,
+                    auth: {
+                        ...state.messages.auth,
+                        changePasswordDone: 'New password has been saved.',
+                    },
+                },
+            };
+        }
+        case AuthActionTypes.CHANGE_PASSWORD_FAILED: {
+            return {
+                ...state,
+                errors: {
+                    ...state.errors,
+                    auth: {
+                        ...state.errors.auth,
+                        changePassword: {
+                            message: 'Could not change password',
+                            reason: action.payload.error.toString(),
+                        },
+                    },
+                },
+            };
+        }
+        case AuthActionTypes.LOAD_AUTH_ACTIONS_FAILED: {
+            return {
+                ...state,
+                errors: {
+                    ...state.errors,
+                    auth: {
+                        ...state.errors.auth,
+                        loadAuthActions: {
+                            message: 'Could not check available auth actions',
                             reason: action.payload.error.toString(),
                         },
                     },